Update sqlex3.c: add casts for MSVC 64-bit warnings

+++ source/sqlex3.c
@@ -312,7 +312,7 @@ HB_BOOL CreateSeekStmt(SQLEXAREAP thiswa, int queryLevel)
 
     pIndexRef = hb_arrayGetItemPtr(thiswa->aOrders, (HB_ULONG)thiswa->hOrdCurrent);
     pColumns = hb_arrayGetItemPtr(pIndexRef, INDEX_FIELDS);
-    thiswa->indexColumns = hb_arrayLen(pColumns);
+    thiswa->indexColumns = (int)hb_arrayLen(pColumns);
 
     // Free the statements we are about to recreate
 
@@ -395,7 +395,7 @@ HB_ERRCODE FeedSeekKeyToBindings(SQLEXAREAP thiswa, PHB_ITEM pKey, int *queryLev
   if (HB_IS_STRING(pKey)) {
     // parse Key string and split it in index fields
 
-    lenKey = hb_itemGetCLen(pKey);
+    lenKey = (int)hb_itemGetCLen(pKey);
     szKey = hb_itemGetCPtr(pKey);
     *queryLevel = thiswa->indexColumns;
